database robot

+++ handlers/app.py
@@ -1,18 +1,27 @@
+import os
+import asyncio
 from os import getenv
 
-from aiogram import Bot, F, Router
+from pathlib import Path
+from aiogram import Bot, F, Router, types
 from aiogram.client.default import DefaultBotProperties
 from aiogram.exceptions import TelegramBadRequest
-from aiogram.filters import CommandStart
-from aiogram.types import Message, CallbackQuery, FSInputFile, InputMediaPhoto, LabeledPrice, KeyboardButton
+from aiogram.filters import CommandStart, Command
+# from aiogram.fsm.context import FSMContext
+# from aiogram.fsm.state import StatesGroup, State
+from aiogram.types import Message, CallbackQuery, FSInputFile, InputMediaPhoto, LabeledPrice, KeyboardButton, ReplyKeyboardMarkup
 from dotenv import load_dotenv
 
-from database.db_utils import *
-from keyboards.inline_kb import *
-from keyboards.reply_kb import *
-from utils.caption import *
+from telegramBotCafe.database.db_utils import *
+from telegramBotCafe.keyboards.inline_kb import *
+from telegramBotCafe.keyboards.reply_kb import *
+from telegramBotCafe.utils.caption import *
+
 
 load_dotenv()
+
+MEDIA_PATH = Path(getenv("MEDIA_PATH")).resolve()
+
 TOKEN = getenv('TOKEN')
 PAY = getenv('PORTMONE')
 MANAGER = getenv('MANAGER')
@@ -21,51 +30,52 @@ bot = Bot(TOKEN, default=DefaultBotProperties(parse_mode='HTML'))
 
 LANGUAGES = {
     "🇬🇧 English": "en",
-    "🇷🇺 Русский": "ru",
     "🇺🇦 Українська": "uk"
 }
-user_language = {}  # TODO Хранилище языков пользователей (можно заменить на базу данных)
+user_language = {}  # TODO Хранилище языков пользователей
 
 
-@router.message(F.text == "⚙️ Настройки")
+def get_image_path(image_name):
+    return MEDIA_PATH / image_name
+
+
+@router.message(CommandStart())
+async def command_start(message: Message):
+    await message.answer(f"Добрий день, <b>{message.from_user.full_name}!</b>\n"
+                         f"вас вітає бот cafesmile!")
+    await start_register_user(message)
+
+
+@router.message(F.text == "⚙️ Налаштування")
 async def start_handler(message: Message):
     buttons = [[KeyboardButton(text=lang)] for lang in LANGUAGES.keys()]
     keyboard = ReplyKeyboardMarkup(keyboard=buttons, resize_keyboard=True, one_time_keyboard=True)
 
-    await message.answer("Выберите язык / Choose your language / Оберіть мову:", reply_markup=keyboard)
+    await message.answer("Оберіть мову / Choose your language:", reply_markup=keyboard)
 
 
 @router.message(F.text.in_(LANGUAGES.keys()))
 async def set_language_handler(message: Message):
     user_id = message.chat.id
-    selected_language = LANGUAGES[message.text]  # Получаем код языка (en, ru, uk)
-    user_language[user_id] = selected_language  # Сохраняем выбор пользователя
+    selected_language = LANGUAGES[message.text]
+    user_language[user_id] = selected_language
 
     if selected_language == "en":
         await message.answer("Language set to English. Let's get started!")
-    elif selected_language == "ru":
-        await message.answer("Язык установлен на русский. Начнем!")
     elif selected_language == "uk":
         await message.answer("Мова встановлена на українську. Розпочнемо!")
 
     await show_main_menu(message)
 
 
-@router.message(CommandStart())
-async def command_start(message: Message):
-    await message.answer(f"Здравствуйте, <b>{message.from_user.full_name}!</b>\n"
-                         f"Вас приветствует бот доставки cafesmile!")
-    await start_register_user(message)
-
-
 async def start_register_user(message: Message):
     chat_id = message.chat.id
     full_name = message.from_user.full_name
     if db_register_user(full_name, chat_id):
-        await message.answer(text='Авторизация прошла успешно')
+        await message.answer(text='Авторизація пройшла успішно')
         await show_main_menu(message)
     else:
-        await message.answer(text='Для связи с Вами нам нужен Ваш контактный номер',
+        await message.answer(text="Для зв'язку з Вами нам потрібний Ваш контактний номер",
                              reply_markup=share_phone_button())
 
 
@@ -75,36 +85,36 @@ async def update_user_info_finish_register(message: Message):
     phone = message.contact.phone_number
     db_update_user(chat_id, phone)
     if db_create_user_cart(chat_id):
-        await message.answer(text="Регистрация прошла успешно!")
+        await message.answer(text="Регістрація пройшла успішно!")
 
     await show_main_menu(message)
 
 
-@router.message(F.text == "🚚 Геолокация и доставка")
+@router.message(F.text == "🎯🏇🏼 Геолокація та доставка")
 async def start_handler(message: Message):
     # Создаем кнопки "Доставка" и "Самовывоз"
-    delivery_button = KeyboardButton(text="🚚 Доставка")
-    pickup_button = KeyboardButton(text="🏠 Самовывоз")
+    delivery_button = KeyboardButton(text="🚛 Доставка")
+    pickup_button = KeyboardButton(text="🚴🏼‍♂️ Самовивіз")
     keyboard = ReplyKeyboardMarkup(keyboard=[[delivery_button, pickup_button]], resize_keyboard=True)
 
-    await message.answer("Выберите способ получения:", reply_markup=keyboard)
+    await message.answer("Виберіть спосіб отримання:", reply_markup=keyboard)
 
 
-@router.message(F.text == "🚚 Доставка")
+@router.message(F.text == "🚛 Доставка")
 async def delivery_handler(message: Message):
-    location_button = KeyboardButton(text="Отправить геолокацию", request_location=True)
+    location_button = KeyboardButton(text="Відправити геолокацію", request_location=True)
     keyboard = ReplyKeyboardMarkup(keyboard=[[location_button]], resize_keyboard=True)
 
-    await message.answer(text="Вы выбрали доставку\n"
-                              "Пожалуйста, отправьте вашу геолокацию для расчета (не забудьте включить местоположение на телефоне)",
+    await message.answer(text="Ви обрали доставку\n"
+                              "Будь ласка, надішліть вашу геолокацію для розрахунку (не забудьте увімкнути місцезнаходження на телефоні)",
                          reply_markup=keyboard)
 
 
-@router.message(F.text == "🏠 Самовывоз")
+@router.message(F.text == "🚴🏼‍♂️ Самовивіз")
 async def pickup_handler(message: Message):
-    await message.answer(text="Вы выбрали самовывоз\n"
-                              "Наш адрес: ул. Примерная, д. 123\n"
-                              "Оплатите пожалуйста покупку, а мы будем вас ждать!",
+    await message.answer(text="Ви обрали самовивіз\n"
+                              "Наша адреса: вул. Мирна, буд. 23\n"
+                              "Оплатіть будь ласка покупку, а ми будемо на вас чекати!",
                          reply_markup=back_to_main_menu())
 
 
@@ -113,25 +123,28 @@ async def location_handler(message: Message):
     latitude = message.location.latitude
     longitude = message.location.longitude
 
-    await message.answer(f"Спасибо! Ваша геолокация получена:\nШирота: {latitude}\nДолгота: {longitude}")
-    await message.answer(text="Доставка будет рассчитана в ближайшее время. Ожидайте!")
-    await message.answer(text="Доставка выходит 100 UAH\n"
-                              "Если вас устраивает, оплатите пожалуйста покупку и ожидайте доставку в течении часа",
+    await message.answer(f"Дякую! Вашу геолокацію отримано:\nШирота: {latitude}\nДовгота: {longitude}")
+    await message.answer(text="Доставка буде розрахована найближчим часом❤️‍🔥Чекайте!")
+    # Ожидание 5 секунд перед следующим сообщением
+    await asyncio.sleep(9)
+    await message.answer(text="Доставка виходить 100 UAH по місту🍾\n"
+                              "Якщо вас влаштовує, сплатіть будь-ласка покупку та чекайте на доставку протягом години🍽\n"
+                              "Для уточнення по замовленню вам може подзвонити наш менеджер🤙\nГарного вам дня та до зустрічі🫰",
                          reply_markup=back_to_main_menu())
 
 
 async def show_main_menu(message: Message):
-    await message.answer(text='Выберите направление',
+    await message.answer(text='🔁 Виберіть напрямок',
                          reply_markup=generate_main_menu())
 
 
-@router.message(F.text == "🎯 Сделать заказ")
+@router.message(F.text == "✅ Зробити замовлення")
 async def make_order(message: Message):
     chat_id = message.chat.id
     await bot.send_message(chat_id=chat_id,
-                           text='Начнем',
+                           text='🪄 Розпочнемо',
                            reply_markup=back_to_main_menu())
-    await message.answer(text="Выберите категорию",
+    await message.answer(text="📋 Виберіть категорію",
                          reply_markup=generate_category_menu(chat_id))
 
 
@@ -145,12 +158,12 @@ async def return_to_main_menu(message: Message):
     await show_main_menu(message)
 
 
-@router.callback_query(F.data.regexp(r'category_[1-9]'))
+@router.callback_query(F.data.regexp(r'category_[1-15]'))
 async def show_product_button(call: CallbackQuery):
     chat_id = call.message.chat.id
     message_id = call.message.message_id
     category_id = int(call.data.split('_')[-1])
-    await bot.edit_message_text(text='Выберите продукт',
+    await bot.edit_message_text(text='🥢 Виберіть продукт',
                                 chat_id=chat_id,
                                 message_id=message_id,
                                 reply_markup=show_product_by_category(category_id))
@@ -162,7 +175,7 @@ async def return_to_category_button(call: CallbackQuery):
     message_id = call.message.message_id
     await bot.edit_message_text(chat_id=chat_id,
                                 message_id=message_id,
-                                text='Выберите категорию',
+                                text='🍎 Виберіть категорію',
                                 reply_markup=generate_category_menu(chat_id))
 
 
@@ -171,31 +184,93 @@ async def show_product_detail(call: CallbackQuery):
     chat_id = call.message.chat.id
     message_id = call.message.message_id
     product_id = int(call.data.split('_')[-1])
+
+    # Получаем продукт из базы данных
     product = db_get_product_by_id(product_id)
+
+    # Проверяем, существует ли товар
+    if not product:
+        await bot.send_message(chat_id=chat_id, text="❌ Помилка: товар не знайдено!")
+        return
+
+    # Получаем категорию через связь product_category
+    category = product.product_category
+    category_id = category.id if category else None  # Получаем id категории
+
+    image_path = get_image_path(product.image)
+
+    # Проверяем, существует ли файл
+    if not os.path.exists(image_path):
+        print(f"Помилка: файл {image_path} не знайдено!")  # Лог в консоли
+        await bot.send_message(chat_id=chat_id, text="Помилка: зображення товару не знайдено!")
+        return
+
+    # class UserState(StatesGroup):
+    #     step1 = State()  # Первый шаг
+    #     step2 = State()  # Второй шаг
+
+    # # Создаем кнопку "Назад"
+    # back_button = InlineKeyboardButton(text="Назад", callback_data="back_to_previous_step")
+
+    # # Создаем разметку для кнопки "Назад" в инлайн-клавиатуре
+    # back_button_markup = InlineKeyboardMarkup(inline_keyboard=[[back_button]])
+
+    # # Удаляем прошлое сообщение
+
     await bot.delete_message(chat_id=chat_id,
                              message_id=message_id)
     if user_cart := db_get_user_cart(chat_id):
         db_update_to_cart(price=product.price, cart_id=user_cart.id)
+        category_id = int(call.data.split('_')[-1])
         text = text_for_caption(product.product_name, product.description, product.price)
         await bot.send_message(chat_id=chat_id,
-                               text='Выберите модификатор',
-                               reply_markup=back_arrow_button())
-
+                               text="🍐 Оберіть товар",
+                               reply_markup=generate_category_menu(chat_id))
         await bot.send_photo(chat_id=chat_id,
-                             photo=FSInputFile(path=product.image),
+                             photo=FSInputFile(str(image_path)),
+                            #  media=InputMediaPhoto(media=FSInputFile(path=product.image), caption=text),
                              caption=text,
-                             reply_markup=generate_constructor_button())
+                             reply_markup=generate_constructor_button(category_id))  # Передаём категорию!
 
     else:
         await bot.send_message(chat_id=chat_id,
-                               text="К сожалению у нас нет Вашего контакта",
+                               text="😿 На жаль, у нас немає Вашого контакту",
                                reply_markup=share_phone_button())
 
 
-@router.message(F.text == '↩️Назад')
-async def return_to_category_menu(message: Message):
-    await bot.delete_message(message.chat.id, message.message_id - 1)
-    await make_order(message)
+# @router.callback_query(F.data == "back_to_previous_step")
+# async def back_to_previous_step(call: CallbackQuery, state: FSMContext):
+#     """ Возвращаем на предыдущий шаг """
+#     user_state = await state.get_state()
+
+#     if user_state == "UserState.step2":
+#         # Переход на шаг 1
+#         await call.message.answer("Вы вернулись к шагу 1.", reply_markup=back_button_markup)
+#         await state.set_state("UserState.step1")  # Переход к шагу 1
+
+#     elif user_state == "UserState.step1":
+#         # Уже на шаге 1, ничего не делаем или отправляем сообщение
+#         await call.message.answer("Вы уже на шаге 1.", reply_markup=back_button_markup)
+
+#     # Ответ на callback_query
+#     await call.answer()
+
+
+# @router.callback_query(F.data == 'back_to_products')
+# async def back_to_products_button(call: CallbackQuery):
+#     chat_id = call.message.chat.id
+    
+#     # Отправляем новое сообщение с выбором товаров или нужной категории
+#     await bot.send_message(
+#         chat_id=chat_id,
+#         text="Оберіть товар",  # Или другой текст, который вам нужен
+#         reply_markup=show_product_by_category(chat_id))  # Функция, возвращающая кнопки для выбора товаров
+
+
+# @router.message(F.text == '↩️Назад')
+# async def return_to_category_menu(message: Message):
+#     await bot.delete_message(message.chat.id, message.message_id - 1)
+#     await make_order(message)
 
 
 @router.callback_query(F.data.regexp(r'action [+-]'))
@@ -217,7 +292,7 @@ async def constructor_change(call: CallbackQuery):
                               cart_id=user_cart.id)
         case '-':
             if user_cart.total_products < 2:
-                await call.answer('Mенее 1 товара выбрать нельзя')
+                await call.answer('❗️ Менше 1 товару вибрати не можна')
             else:
                 user_cart.total_products -= 1
             product_price = product_price * user_cart.total_products
@@ -251,22 +326,22 @@ async def put_info_cart(call: CallbackQuery):
                                   total_products=cart.total_products,
                                   total_price=cart.total_price):
         await bot.send_message(chat_id=chat_id,
-                               text='Продукт успешно добавлен ☑️')
+                               text='☑️ Продукт успішно доданий')
     else:
         await bot.send_message(chat_id=chat_id,
-                               text='Количество успешно изменено ✏️')
+                               text='✏️ Кількість успішно змінено')
 
-    await return_to_category_menu(call.message)
+    await generate_category_menu(call.message)
 
 
-@router.callback_query(F.data == 'Ваша корзина')
+@router.callback_query(F.data == 'Ваш кошик')
 async def show_finally_cart(call: CallbackQuery):
     """Показ содержимого корзины"""
     message_id = call.message.message_id
     chat_id = call.from_user.id
     await bot.delete_message(chat_id=chat_id,
                              message_id=call.message.message_id)
-    context = counting_products_from_cart(chat_id, 'Ваша корзина:')
+    context = counting_products_from_cart(chat_id, '🧺 Ваш кошик:')
     if context:
         count, text, *_ = context
         await bot.send_message(chat_id=chat_id,
@@ -274,7 +349,7 @@ async def show_finally_cart(call: CallbackQuery):
                                reply_markup=generate_delete_product(chat_id))
     else:
         await bot.send_message(chat_id=chat_id,
-                               text='Ваша корзина пуста😱')
+                               text='😱 Ваш кошик порожній')
         await make_order(call.message)
 
 
@@ -284,7 +359,7 @@ async def delete_cart_product(call: CallbackQuery):
     finally_id = int(call.data.split('_')[-1])
     db_delete_product(finally_id)
     await bot.answer_callback_query(callback_query_id=call.id,
-                                    text='Продукт удален с корзины!')
+                                    text='Продукт видалено з кошика❗️')
     await show_finally_cart(call)
 
 
@@ -297,48 +372,56 @@ async def create_order(call: CallbackQuery):
                              message_id=message_id)
 
     count, text, total_price, cart_id = counting_products_from_cart(chat_id=chat_id,
-                                                                    user_text='Оплата заказа:')
-    text += "\nДоставка по городу 100 UAH"
+                                                                    user_text='🧮 Оплата замовлення:')
+    text += "\n💰 Доставка по місту 100 UAH"
 
     await bot.send_invoice(chat_id=chat_id,
-                           title='Ваш заказ:',
+                           title='Ваше замовлення:',
                            description=text,
                            payload='bot-defined invoice payload',
                            provider_token=PAY,
                            currency='UAH',
                            prices=[
-                               LabeledPrice(label='Общая стоимость', amount=int(total_price) * 100),
+                               LabeledPrice(label='🧮 Загальна вартість', amount=int(total_price) * 100),
                                LabeledPrice(label='Доставка', amount=100 * 100)
                            ])
 
-    await bot.send_message(chat_id=chat_id,
-                           text='Заказ оплачен. Спасибо!')
+    # Отправляем отчет менеджеру (без очистки корзины на этом этапе)
     await sending_report_to_manager(chat_id, text)
 
-    clear_finally_cart(cart_id)
+
+@router.message(F.successful_payment)
+async def payment_successful(message: Message):
+    """Обработчик успешной оплаты"""
+    chat_id = message.from_user.id
+
+    try:
+        clear_finally_cart(chat_id)  # Очищаем корзину по ID пользователя
+        await message.answer("Дякую за вашу оплату❗️ Замовлення оформлене, кошик очищений")
+    except Exception as e:
+        await message.answer(f"Помилка під час очищення кошика: {e}")
 
 
 async def sending_report_to_manager(chat_id: int, text: str):
     """Отправка заказов в чат группы"""
     user = db_get_user_info(chat_id)
-    text += f'\n\nИмя клиента: {user.name}\nКонтакт: {user.phone}\n\n'
+    text += f"\n\nІм'я клієнта: {user.name}\nКонтакт: {user.phone}\n\n"
     await bot.send_message(chat_id=MANAGER,
                            text=text)
 
 
-@router.message(F.text == '🧺 Корзина')
+@router.message(F.text == '🧺 Кошик')
 async def make_order(message: Message):
     chat_id = message.chat.id
     await bot.send_message(chat_id=chat_id,
-                           text='Начнем',
+                           text='Почнемо',
                            reply_markup=back_to_main_menu())
-    context = counting_products_from_cart(chat_id, 'Ваша корзина:')
+    context = counting_products_from_cart(chat_id, '🧺 Ваш кошик:')
     if context:
         count, text, *_ = context
         await bot.send_message(chat_id=chat_id,
                                text=text,
                                reply_markup=generate_delete_product(chat_id))
     else:
-        await bot.send_message(chat_id=chat_id,
-                               text='Ваша корзина пуста😱')
-        await make_order(message)
+        await bot.send_message(chat_id=chat_id, text='😱 Ваш кошик порожній')
+
